Refactor Data category/element and master clause

+++ gdpr/src/main/java/com/kairos/controller/data_category_element/DataCategoryController.java
@@ -41,7 +41,7 @@ public class DataCategoryController {
 
     @ApiOperation("get data category by id with data Elements ")
     @GetMapping(COUNTRY_URL + "/data_category/{dataCategoryId}")
-    public ResponseEntity<ResponseDTO<DataCategoryResponseDTO>> getDataCategoryWithDataElementsById(@PathVariable Long countryId, @PathVariable BigInteger dataCategoryId) {
+    public ResponseEntity<ResponseDTO<DataCategoryResponseDTO>> getDataCategoryWithDataElementsById(@PathVariable Long countryId, @PathVariable Long dataCategoryId) {
         return ResponseHandler.generateResponseDTO(HttpStatus.OK, true, dataCategoryService.getDataCategoryWithDataElementByCountryIdAndId(countryId, dataCategoryId));
 
     }
@@ -62,7 +62,7 @@ public class DataCategoryController {
 
     @ApiOperation("update data category by id ")
     @PutMapping(COUNTRY_URL + "/data_category/{dataCategoryId}")
-    public ResponseEntity<ResponseDTO<DataCategoryDTO>> updateDataCategoryAndDataElement(@PathVariable Long countryId, @PathVariable BigInteger dataCategoryId, @Valid @RequestBody DataCategoryDTO dataCategoryDto) {
+    public ResponseEntity<ResponseDTO<DataCategoryDTO>> updateDataCategoryAndDataElement(@PathVariable Long countryId, @PathVariable Long dataCategoryId, @Valid @RequestBody DataCategoryDTO dataCategoryDto) {
         return ResponseHandler.generateResponseDTO(HttpStatus.OK, true, dataCategoryService.updateDataCategoryAndDataElement(countryId, false, dataCategoryId, dataCategoryDto));
 
     }
@@ -76,14 +76,14 @@ public class DataCategoryController {
 
     @ApiOperation("organization update data category by id ")
     @PutMapping(UNIT_URL + "/data_category/{dataCategoryId}")
-    public ResponseEntity<ResponseDTO<DataCategoryDTO>> updateOrganizationDataCategoryAndDataElement(@PathVariable Long unitId, @PathVariable BigInteger dataCategoryId, @Valid @RequestBody DataCategoryDTO dataCategoryDto) {
+    public ResponseEntity<ResponseDTO<DataCategoryDTO>> updateOrganizationDataCategoryAndDataElement(@PathVariable Long unitId, @PathVariable Long dataCategoryId, @Valid @RequestBody DataCategoryDTO dataCategoryDto) {
         return ResponseHandler.generateResponseDTO(HttpStatus.OK, true, dataCategoryService.updateDataCategoryAndDataElement(unitId, true, dataCategoryId, dataCategoryDto));
 
     }
 
     @ApiOperation("organization, get data category by id with data Elements ")
     @GetMapping(UNIT_URL + "/data_category/{dataCategoryId}")
-    public ResponseEntity<ResponseDTO<DataCategoryResponseDTO>> getOrganizationDataCategoryWithDataElements(@PathVariable Long unitId, @PathVariable BigInteger dataCategoryId) {
+    public ResponseEntity<ResponseDTO<DataCategoryResponseDTO>> getOrganizationDataCategoryWithDataElements(@PathVariable Long unitId, @PathVariable Long dataCategoryId) {
         return ResponseHandler.generateResponseDTO(HttpStatus.OK, true, dataCategoryService.getDataCategoryWithDataElementByUnitIdAndId(unitId, dataCategoryId));
 
     }
